Restructured code into modules as the original file was getting too large
